+++ packages/common-server/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./files";
 export * from "./filesv2";
 export * from "./logger";
-export * from "./api";
 export * from "./parser";
 export * from "./git";
 export * from "./etc";