+++ packages/engine-server/src/markdown/remark/wikiLinks.ts
@@ -16,7 +16,7 @@ import {
   DendronASTTypes,
 } from "../types";
 import { MDUtilsV4 } from "../utils";
-import { MDUtilsV5 } from "../utilsv5";
+import { MDUtilsV5, ProcMode } from "../utilsv5";
 import { addError, getNoteOrError, LinkUtils } from "./utils";
 
 export const LINK_REGEX = /^\[\[(.+?)\]\]/;
@@ -82,7 +82,7 @@ function attachCompiler(proc: Unified.Processor, opts?: CompilerOpts) {
       const data = node.data;
       let value = node.value;
 
-      if (pOpts.useRaw) {
+      if (pOpts.mode === ProcMode.NO_DATA) {
         const { alias, anchorHeader } = data;
         let link = value;
         let calias = alias !== value ? `${alias}|` : "";
@@ -184,11 +184,11 @@ function attachParser(proc: Unified.Processor) {
     if (_.isNull(out)) {
       throw new DendronError({ message: `link is null: ${linkMatch}` });
     }
-    if (pOpts.useRaw) {
+    if (pOpts.mode === ProcMode.NO_DATA) {
       return out;
     }
 
-    let { config, vault, dest, fname } = MDUtilsV4.getDendronData(proc);
+    let { config, vault, dest, fname } = MDUtilsV5.getProcData(proc);
     const { engine } = MDUtilsV4.getEngineFromProc(proc);
     if (out.vaultName) {
       const maybeVault = VaultUtils.getVaultByName({
@@ -210,6 +210,7 @@ function attachParser(proc: Unified.Processor) {
       // default to current note
     }
     if (!out.value) {
+      debugger;
       // same file block reference, value is implicitly current file
       out.value = _.trim(NoteUtils.normalizeFname(fname)); // recreate what value (and alias) would have been parsed
       if (!out.alias) out.alias = out.value;